backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
blis==1.3.3
boto3==1.42.90
botocore==1.42.90
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
chardet==6.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cobble==0.1.4
colorama==0.4.6
confection==1.3.3
coverage==7.13.4
cryptography==46.0.5
cymem==2.0.13
distro==1.9.0
dnspython==2.8.0
easyocr==1.7.2
ecdsa==0.19.2
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
fastapi==0.129.0
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.64.0
grpcio==1.78.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.14
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.6
lazy-loader==0.5
lxml==6.0.2
mammoth==1.11.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
networkx==3.6.1
ninja==1.13.0
nltk==3.9.4
numpy==2.4.2
openai==2.21.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
packaging==26.0
pandas==3.0.2
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
portalocker==3.2.0
preshed==3.0.13
propcache==0.4.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
PyMuPDF==1.27.1
pypdf==6.10.0
pytest==9.0.2
pytest-cov==7.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
tifffile==2026.4.11
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wasabi==1.1.3
weasel==1.0.0
websockets==15.0.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
